+++ cypress/integration/tools/VisitHomePage.js
@@ -1,1 +1,12 @@
-require('../../support/TestListManager').AddToCypressTestList('Tools.VisitHomePage')
+/**
+* Copyright (c) Microsoft Corporation. All rights reserved.  
+ * Licensed under the MIT License.
+*/
+
+const homePage = require('../../support/components/HomePage')
+
+export function VisitHomePage()
+{
+  homePage.Visit()
+  homePage.GetModelListRowCount()
+}
